requirements.txt
supabase>=2.0.0
python-dotenv>=1.0.0
whatsapp-web.py>=0.0.1
qrcode-terminal>=0.1
